+++ cookbooks/chub_mc_app/recipes/default.rb
@@ -61,6 +61,14 @@ service "#{node[:chub_mc_app][:app_name]}" do
     action [ "disable", "stop" ]
 end
 
+file "#{node[:chub_mc_app][:staging_dir]}/#{node[:chub_mc_app][:app_name]}.jar" do
+    action   :delete
+    mode     "0755"
+    owner    "chub_#{node[:chub_mc_app][:app_name]}"
+    group    "chub_#{node[:chub_mc_app][:app_name]}"
+end
+
+
 remote_file "#{node[:chub_mc_app][:staging_dir]}/#{node[:chub_mc_app][:app_name]}.jar" do
   source "#{node[:chub_mc_app][:jar_file_url]}"
   owner "chub_#{node[:chub_mc_app][:app_name]}"
@@ -78,8 +86,8 @@ end
 remote_file "Copy deploy jar file from staging" do 
   path "#{node[:chub_mc_app][:deploy_dir]}/#{node[:chub_mc_app][:app_name]}.jar"
   source "file://#{node[:chub_mc_app][:staging_dir]}/#{node[:chub_mc_app][:app_name]}.jar"
-  owner 'chub_#{node[:chub_mc_app][:app_name]}'
-  group 'chub_#{node[:chub_mc_app][:app_name]}'
+  owner "chub_#{node[:chub_mc_app][:app_name]}"
+  group "chub_#{node[:chub_mc_app][:app_name]}"
   mode 0755
 end