+++ cookbooks/chub_plaza/environments/ssodev.rb
@@ -13,5 +13,8 @@ default_attributes(
     :cas_service_url => "http://plaza.vagrant.dev:8080/shiro-cas",
     :cas_failure_url => "http://plaza.vagrant.dev:8080/auth/fail",
     :census_url => "https://ssodev1-census.nexus.commercehub.com:8443"
+    :cas_failure_url => "http://plaza.vagrant.dev:8080/auth/fail",
+    :census_url => "https://ssodev1-census.nexus.commercehub.com:8443",
+    :vault_url => "https://ssodev1-vault.nexus.commercehub.com:8443"
   }
 )