+++ environments/ssoprod1.rb
@@ -3,7 +3,7 @@ description "SSO Prod environment 1"
 cookbook_versions({
   "apache2" => "= 1.7.0",
   "apt" => "= 2.3.8",
-  "base" => "= 0.1.15",
+  "base" => "= 0.1.16",
   "bluepill" => "= 2.3.1",
   "build-essential" => "= 1.4.2",
   "chef-client" => "= 3.0.6",
@@ -43,6 +43,10 @@ cookbook_versions({
 })
 default_attributes(
   :sensu => {
+    :api => {
+      :user => "sensuapi",
+      :password => "@HrMgi1I3exo"
+    },
     :rabbitmq => {
       :host => "ssoprod1sensu.commercehub.com"
     }
@@ -100,6 +104,7 @@ default_attributes(
   },
   :chub_castle => {
     :app_url => "file:///var/castle/staged-castle.war",
+    :domain => "castle.commercehub.com",
     :default_service_url => "https://plaza.commercehub.com/shiro-cas",
     :forgot_password_url => "https://plaza.commercehub.com/forgotPassword",
     :vault_url => "https://ssoprod1-vault.commercehub.com:8443",
@@ -109,7 +114,7 @@ default_attributes(
       :password => "ssoprod-pass",
       :multicast_group => "224.2.2.6",
       :multicast_port => 54330,
-      :service_ticket_timeout => 300,
+      :service_ticket_timeout => 1800,
       :ticket_granting_ticket_timeout => 7200
     }
   },
@@ -118,6 +123,7 @@ default_attributes(
     :cas_server_url => "https://castle.commercehub.com",
     :cas_service_url => "https://plaza.commercehub.com/shiro-cas",
     :cas_failure_url => "https://plaza.commercehub.com/",
-    :census_url => "https://ssoprod1-census.commercehub.com:8443"
+    :census_url => "https://ssoprod1-census.commercehub.com:8443",
+    :vault_url => "https://ssoprod1-vault.commercehub.com:8443"
   }
 )