+++ roles/buyspace_wapp.rb
@@ -1,0 +1,12 @@
+name "buyspace_wapp"
+description "role defining buyspace web app node"
+default_attributes "chub-buyspace" => {
+    "taskScheduler" => {
+      "enabled" => false
+    }, 
+    "grails" => {
+      "serverURL" => "https://mpqa1-buyspace.commercehub.com",
+      "apiServerURL" => "http://mpqa1-api.commercehub.com"
+    }   
+}
+run_list "role[base]", "recipe[chub-buyspace::upgrade_buyspace]"